chore: put WASM interface in its own crate
