+++ wonnx-wasm/index.html
@@ -5,7 +5,7 @@
 	</head>
 
 	<body>
-		<h1>WONNX</h1>
+		<h1>WONNX on the web</h1>
 		<textarea id="log" style="width: 100%; height: 300px;"></textarea>
 		
 		<script type="module">
@@ -24,21 +24,17 @@
 
 			async function run() {
 				try {
-					const [modelBytes, initResult] = await Promise.all([fetchBytes("models/single_relu.onnx"), init()])
+					const [modelBytes, initResult] = await Promise.all([fetchBytes("../data/models/single_relu.onnx"), init()])
 					console.log("Initialized", { modelBytes, initResult, Session});
 					const session = await Session.fromBytes(modelBytes);
 
 					
 					console.log({session});
-					let result = null;
-
-					for(let i = 0; i < 1000; i++) {
-						const input = new Input();
-						input.insert("x", [13.0, -37.0]);
-						console.time("inference");
-						result = await session.run(input);
-						console.timeEnd("inference");
-					}
+					const input = new Input();
+					input.insert("x", [13.0, -37.0]);
+					console.time("inference");
+					const result = await session.run(input);
+					console.timeEnd("inference");
 					console.log({result});
 					log(JSON.stringify(result, undefined, "\t"));
